+++ x-pack/test/api_integration/apis/upgrade_assistant/upgrade_assistant.ts
@@ -14,8 +14,7 @@ export default function ({ getService }: FtrProviderContext) {
   const es = getService('es');
   const supertest = getService('supertest');
 
-  // Failing: See https://github.com/elastic/kibana/issues/199719
-  describe.skip('Upgrade Assistant', function () {
+  describe('Upgrade Assistant', function () {
     describe('Reindex operation saved object', () => {
       const dotKibanaIndex = '.kibana';
       const fakeSavedObjectId = 'fakeSavedObjectId';
